Consolidate `TickType` enum
This one definition does not warrant its own package hierarchy.

Helps to slim down `trinity.components.eth2`
